src/requirements.txt
asgiref==3.8.1
astroid==3.1.0
dill==0.3.8
Django==5.0.3
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
isort==5.13.2
Markdown==3.6
mccabe==0.7.0
platformdirs==4.2.0
pylint==3.1.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.4
typing_extensions==4.10.0
